requirements.txt
anthropic>=0.40.0
openai>=1.40.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
rich>=13.0.0
